textassistantchat/requirements.txt
torch
transformers
streamlit
chromadb
sentence-transformers
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-nvidia-ai-endpoints